there is no "long long" type in C++
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@2277 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Base/TextBase.cpp
@@ -361,7 +361,7 @@ void TextBase :: makeLineDist()
     {
       // now in the other direction:
       // [-n ... -2 -1 0]
-	  signed long long j;
+      signed long j;
       for(j=m_theText.size()-1; j>=0; j--){
         m_lineDist.push_back(-j);
       }
